+++ src/CardText.java
@@ -1,9 +1,9 @@
 public class CardText {
     // This object stores both the text of an effect and its cost. It is used instead of storing these as an array.
     private String text;
-    private Integer cost;
+    private double cost;
 
-    public CardText(String text, Integer cost) {
+    public CardText(String text, double cost) {
         this.text = text;
         this.cost = cost;
     }
@@ -12,5 +12,5 @@ public class CardText {
     *that bad*. I'm just afraid that my Java 101 teacher will somehow find this and steal my degree away from me. */
     public String getText() { return text; }
 
-    public Integer getCost() { return cost; }
+    public double getCost() { return cost; }
 }
